+++ src/ellipsoid.h
@@ -1,5 +1,16 @@
-#ifndef __ellipsoid_h__
-#define __ellipsoid_h__
+#pragma once
+
+/* Utility functions to handle ellipsoidal geometry
+ *
+ * References:
+ * Feltens J (2009) Vector method to compute the Cartesian (X, Y , Z)
+ * to geodetic (, λ, h) transformation on a triaxial ellipsoid. J
+ * Geod 83:129–137 10.1007/s00190-008-0246-5
+ *
+ * Sebahattin Bektas (2014), Orthogonal distance from an ellipsoid,
+ * http://dx.doi.org/10.1590/S1982-21702014000400053
+ */
+
 
 #include <math.h>
 #include <assert.h>
@@ -19,12 +30,6 @@ typedef struct{
   double eps;
 } elli;
 
-/*
- * Feltens    J    (2009)    Vector    method    to compute   the   Cartesian   (X, Y , Z)   to geodetic   (,  λ, h)   transformation   on   a triaxial ellipsoid. J Geod 83:129–137
- * 10.1007/s00190-008-0246-5
- *
- * Bektas: http://dx.doi.org/10.1590/S1982-21702014000400053 
- */
 
 /* Create a new ellipsoid. Can be free'd with free */
 elli * elli_new(double a, double b, double c);
@@ -32,14 +37,16 @@ elli * elli_new(double a, double b, double c);
 /* Print it */
 void elli_show(elli * E);
 
+/*  */
 double elli_dist(elli * E, double * P);
-// Volume
+
+/* Volume */
 double elli_vol(elli * E);
 
-// 1 if point p is inside E, 0 otherwise
+/* 1 if point p is inside E, 0 otherwise */
 int elli_isInside(elli * E, double * p);
 
-// Distance to surface
+/* Distance to surface */
 double elli_dist(elli * E, double * p);
 
 /* Generalized radius
@@ -48,16 +55,22 @@ double elli_dist(elli * E, double * p);
  */
 double elli_radius(elli * E, double * p);
 
-/* Geodesic (=shortest) distance from P to ellipse 
+/* Geodesic (=shortest) distance from P to ellipse
  * if Y != NULL, it will return the closest point */
 double elli_gdist(elli * ellipse, double * P, double * Y);
-// Using Lagrange multiplier approach
-double elli_gdistL(elli * ellipse, double * P, double * Y);
+
+/* Using Lagrange multiplier approach */
+double elli_gdistL(elli * ellipse,
+                   const double * P,
+                   double * Y);
 
 /* get scaling parameter but do not appy it to X */
-double elli_getScale(elli * E, double * X);
-// Same as above but squared (i.e. not sqrt -- faster)
-double elli_getScale2(elli * E, double * X);
+double elli_getScale(const elli * restrict E,
+                     const double * restrict X);
+
+/* Same as above but squared (i.e. not sqrt -- faster) */
+double elli_getScale2(const elli * restrict E,
+                      const double * restrict X);
 
 /* Scale X so that it is on E */
 void elli_scale(elli * E, double * X);
@@ -66,7 +79,6 @@ void elli_scale(elli * E, double * X);
  * Undefined behaviour if P isn't on E
  */
 void elli_normal(elli * E, double * P, double * N);
-// Normalize N to 1
-void elli_nnormal(elli * E, double * P, double * N);
 
-#endif
+/* Normalize N to 1 */
+void elli_nnormal(elli * E, double * P, double * N);