+++ src/mflock_cli.c
@@ -1,0 +1,6 @@
+#include "mflock.h"
+
+int main(int argc, char ** argv)
+{
+    return mflock(argc, argv);
+}